🚧 Experimenting with the spectra

+++ src/args.rs
@@ -43,8 +43,8 @@ pub struct Args {
 #[derive(Parser)]
 pub struct TracerOptions {
     /// Samples per pixel that get averaged for the antialiasing.
-    #[arg(short = 's', long = "samples", default_value = "1", value_parser = value_parser!(u16).range(1..))]
-    pub samples_per_pixel: usize,
+    #[arg(short = 's', long = "samples", default_value = "1", value_parser = value_parser!(u32).range(1..))]
+    pub samples_per_pixel: u32,
 
     /// Maximum number of ray bounces of the scene's surfaces.
     #[arg(short = 'b', long = "max-bounces", default_value = "5", value_parser = value_parser!(u16).range(1..))]
